test: added test to check user is informed (exception TableAlreadyExistsException is thrown) if on addTable the specified table already exists. feat: implemented check in addTable for existing table

+++ src/main/java/uk/co/imperatives/exercise/repository/TableRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.repository.CrudRepository;
 import uk.co.imperatives.exercise.model.Guest;
 import uk.co.imperatives.exercise.model.Table;
 
-public interface TableRepository extends CrudRepository<Table, Long> {
+public interface TableRepository extends CrudRepository<Table, Integer> {
 }
